Add a non-optimized non-incremental Bazaar loader
The work for optimizing the loader has been paved with issues, in
the mean time, this works and seems robust enough on a test run
on all launchpad projects.

The incremental logic will be added in a future patch.

This also updates the requirements file to more explicitly match newer
versions of the swh dependencies.