Support new repositories with old branches
You can have a branch that does not support tags located in a repository
that has the newest format. This fixes T3963.

The check that the upgrade is actually done is duplicated
in `test_needs_upgrade`, since the test would pass fine with an already
upgraded repo, albeit with a lower coverage.