loader: Add parameter to check validity of computed revisions
Add check_revision int parameter to the BzrLoader class.

Similarly to the subversion loader, every check_revision computed revisions
the filesystem associated to a bazaar revision is exported to a temporary
folder, the hash of its root directory computed and then compared to the SWH
directory id associated to the SWH revision. If they differ, the loading is
stopped by raising an exception.

Use this new parameter in tests to check sample repositories are correctly
loaded into the archive.