loader: Do not include content of nested trees
Previously the loader was raising an error when encountering a nested
tree (inventory entry of kind 'tree-reference').

Those nested trees are typically created from git submodules when importing
a bazaar repository from a git one.

By default bazaar does not recurse on a nested tree content when exporting
a revision and simply creates an empty directory instead.

So adopt the same behavior in the loader as it exists cases where a nested
tree cannot be resolved.

Related to #4583.